fixed a MSVC warning; corrected a typo
Former-commit-id: f51b5d6f8301b2f5a4c46c4917b236e8231f4423

+++ src/Codecs/MappedRecord.cc
@@ -34,14 +34,14 @@ MappedRecord::MappedRecord(const SAMRecord &samRecord, const uint32_t &positionO
         case 'M':
         case '=':
         case 'X':
-            lastPosition += opLen;
+            lastPosition += (uint32_t)opLen;
             break;
         case 'I':
         case 'S':
             break;
         case 'D':
         case 'N':
-            lastPosition += opLen;
+            lastPosition += (uint32_t)opLen;
             break;
         case 'H':
         case 'P':
